requirements.txt
fastapi
uvicorn
pandas
google-generativeai
gspread
oauth2client
python-dotenv
openpyxl
